implementacion de logica

+++ src/Library/StatsChangerMove.cs
@@ -4,6 +4,6 @@ public class StatsChangerMove: IMove
 {
     public string Name { get; }
     public bool Special { get; }
-    public double Buff { get; }
+    public int Buff { get; }
     public int LastTurnPlayed { get; }
 }
